refactor: split MLA latent CP implementation
